Move apiEnvFlags to env.ts - #2126
