+++ src/apiDefs/env.ts
@@ -6,3 +6,19 @@ export function isHostedApiEnabled(shortName: string, defaultValue: boolean): bo
     return envValue === 'true';
   }
 }
+
+export const apiEnvFlags = {
+  address_validation: isHostedApiEnabled('address_validation', true),
+  appeals: isHostedApiEnabled('appeals', true),
+  argonaut: isHostedApiEnabled('argonaut', true),
+  benefits: isHostedApiEnabled('benefits', true),
+  claims: isHostedApiEnabled('claims', true),
+  community_care: isHostedApiEnabled('community_care', true),
+  disability_rating: isHostedApiEnabled('disability_rating', true),
+  facilities: isHostedApiEnabled('facilities', true),
+  fhir: isHostedApiEnabled('fhir', true),
+  loan_guaranty: isHostedApiEnabled('loan_guaranty', false),
+  service_history: isHostedApiEnabled('service_history', true),
+  urgent_care: isHostedApiEnabled('fhir', true),
+  veteran_confirmation: isHostedApiEnabled('veteran_confirmation', true),
+};
