+++ src/apiDefs/index.ts
@@ -13,7 +13,6 @@ import {
 } from '../content/releaseNotes';
 
 import benefitsApis from './benefits';
-import { isHostedApiEnabled } from './env';
 import facilitiesApis from './facilities';
 import healthApis from './health';
 import verificationApis from './verification';
@@ -165,19 +164,3 @@ function apisFor(apiList: string[]): IApiDescription[] {
 export function includesOauthAPI(apiList: string[]): boolean {
   return categoriesFor(apiList).some(category => !category.apiKey)|| apisFor(apiList).some(api => api.oAuth || false);
 }
-
-export const apiEnvFlags = {
-  address_validation: isHostedApiEnabled('address_validation', true),
-  appeals: isHostedApiEnabled('appeals', true),
-  argonaut: isHostedApiEnabled('argonaut', true),
-  benefits: isHostedApiEnabled('benefits', true),
-  claims: isHostedApiEnabled('claims', true),
-  community_care: isHostedApiEnabled('community_care', true),
-  disability_rating: isHostedApiEnabled('disability_rating', true),
-  facilities: isHostedApiEnabled('facilities', true),
-  fhir: isHostedApiEnabled('fhir', true),
-  loan_guaranty: isHostedApiEnabled('loan_guaranty', false),
-  service_history: isHostedApiEnabled('service_history', true),
-  urgent_care: isHostedApiEnabled('fhir', true),
-  veteran_confirmation: isHostedApiEnabled('veteran_confirmation', true),
-};
